Refactor StorageModuleException
