+++ src/main/java/seedu/duke/constants/ExceptionMessages.java
@@ -21,5 +21,6 @@ public class ExceptionMessages {
     public static final String MESSAGE_DUPLICATE_MODULE_PARAMETERS = "Duplicate parameters."
             + "Please follow the syntax module add c/MODULE_CODE {n/[MODULE_NAME] "
             + "m/[MODULAR CREDITS] e/[EXPECTED_GRADE]}";
-
+    public static final String MESSAGE_STORAGE_EXCEPTION = "Something is wrong with the Module storage file. "
+            + "A new file has been created.";
 }
