+++ src/test/java/seedu/duke/parser/module/ParserModuleTest.java
@@ -1,7 +1,7 @@
 package seedu.duke.parser.module;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.commands.module.StorageModuleException;
+import seedu.duke.exceptions.module.StorageModuleException;
 import seedu.duke.exceptions.storage.StorageException;
 import seedu.duke.module.Module;
